Refactor functionality into handler class to allow sharing between spell creation fragment and dialog.
